+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.1] - 2026-09-24
+
+### Fixed
+
+- Exit with an error when a given path does not exist or contains no Python
+  files, instead of passing without checking anything (e.g. after a package
+  directory is renamed).
+
 ## [0.1.0] - 2026-09-24
 
 ### Added
@@ -19,5 +27,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     SQLModel `Relationship(...)` attributes.
   - `--exclude` and `--model-base` options.
 
-[Unreleased]: https://github.com/amirzenoozi/pre-commit-hooks/compare/v0.1.0...HEAD
+[Unreleased]: https://github.com/amirzenoozi/pre-commit-hooks/compare/v0.1.1...HEAD
+[0.1.1]: https://github.com/amirzenoozi/pre-commit-hooks/compare/v0.1.0...v0.1.1
 [0.1.0]: https://github.com/amirzenoozi/pre-commit-hooks/releases/tag/v0.1.0